backend-requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pillow==10.1.0
firebase-admin==6.2.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
django-filter==23.4
drf-spectacular==0.26.4
gunicorn==21.2.0
whitenoise==6.6.0
